Refactors repair-shop module types and enums
Moves repair-shop related type definitions and enums from application-specific paths to a centralized `src/modules/repair-shop` directory. This improves modularity and organization.

Renames ORM-related interfaces with an `ORM` suffix for clearer distinction between database models and other data structures.
