--- src/app/(auth)/repair-shop/sales/_parts/types/sale-service.ts
@@ -1,9 +1,0 @@
-import type Service from '@/app/(auth)/repair-shop/services/_parts/types/service'
-
-export default interface SaleService {
-    id: number
-    sale_uuid: string
-    service_id: string
-    rp: number
-    state: Service
-}
