+++ src/modules/repair-shop/enums/vehicle-type.ts
